% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFEmetab.R
\name{calcZMixDens}
\alias{calcZMixDens}
\title{Calculate Z Mix Density
Using density instead of temp to indicate mixed layer, (though note that objects in this script are still labeled e.g. 'temps', not 'dens')}
\usage{
calcZMixDens(dataIn, thresh = 0.075)
}
\arguments{
\item{dataIn}{A density profile. First column is datetime, POSIXct. Second and subsequent columns are density (kg m-3) at depth. Column headers for these columns are e.g. temp0, temp0.5, temp1.0, temp2, temp2.3}

\item{thresh}{Threshold density change to indicate end of mixed layer. Units (kg m-3) m-1. Defaults to 0.075 (value Jordan Read is using in their project)}
}
\value{
A data frame. First column is datetime, second column is zMix calculated at that datetime.
}
\description{
Calculate Z Mix Density
Using density instead of temp to indicate mixed layer, (though note that objects in this script are still labeled e.g. 'temps', not 'dens')
}
