% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfuns.R
\name{checkCols}
\alias{checkCols}
\title{Check for missing/extra columns}
\usage{
checkCols(old, new)
}
\arguments{
\item{old}{A data frame (the old/original version)}

\item{new}{A data frame (the new version)}
}
\value{
Gives warnings/messages to tell about missing or extra columns.
}
\description{
Check for missing/extra columns
}
