% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{checkDistance}
\alias{checkDistance}
\title{Check Distance
Function to check the distance between pairs}
\usage{
checkDistance(df, threshold = 1)
}
\arguments{
\item{df}{Data frame with (at least) columns "DOC" and "sampleID"; otherwise this will throw an error.}

\item{threshold}{Specifies the threshold value for classifying a pair of DOC replicates as "far" apart or not. Default is 1: e.g. points are "far" if the higher one is more than 100\% greater than the lower one. If you set a threshold over 2, the function will warn you that this threshold may be unreasonable/not helpful, but it won't stop you from setting the threshold to whatever you want.}
}
\value{
A data frame containing sampleID's and the result of the near/far check, only.
}
\description{
Check Distance
Function to check the distance between pairs
}
