% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfuns.R
\name{checkINFO}
\alias{checkINFO}
\title{Check that all INFO ID's are in SAMPLES}
\usage{
checkINFO(info, sample)
}
\arguments{
\item{info}{A data frame, e.g. FISH_INFO.}

\item{sample}{A data frame, e.g. FISH_SAMPLES.}
}
\value{
Messages about whether or not all the ID's from INFO are also present in SAMPLES.
}
\description{
Check that all INFO ID's are in SAMPLES
}
