% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfuns.R
\name{checkMet}
\alias{checkMet}
\title{Check that all metadataID's show up in the METADATA table. Requires that you have previously connected to the database using dbTable().}
\usage{
checkMet(new, old = dbTable("METADATA")$metadataID)
}
\arguments{
\item{new}{A data frame (the old/original version)}
}
\value{
Gives warnings/messages about metadataID's and whether they show up in the METADATA table.
}
\description{
Check that all metadataID's show up in the METADATA table. Requires that you have previously connected to the database using dbTable().
}
