% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{colsCheck}
\alias{colsCheck}
\title{Check that a data frame has certain columns}
\usage{
colsCheck(obj, cols)
}
\arguments{
\item{obj}{An object that you want to check for cols}

\item{cols}{A vector of column names you want to check for}
}
\value{
Throws an informative error if any of \code{cols} are missing from \code{obj}. Additionally throws an error if \code{obj} is not a data frame (see \code{dfCheck()})
}
\description{
\code{colsCheck} checks whether a data frame contains the specified columns. You can check for as many columns as you want, and the function will throw an informative error if \emph{any} of them are missing (and it will tell you which ones) are missing. It will also check that \code{obj} is a data frame.
}
