% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{dateCheck}
\alias{dateCheck}
\title{Check \code{sampleID} date against \code{dateSample}}
\usage{
dateCheck(df)
}
\arguments{
\item{df}{A data frame to check}
}
\value{
ID's of rows that have mismatched dates.
}
\description{
\code{dateCheck} checks whether all the dates in the \code{sampleID} column match up with the dates in the \code{dateSample} column. If they don't, returns numeric indices of the rows that contain mismatches. This function additionally verifies that \code{df} is a data frame and that it has the columns \code{sampleID} and \code{dateSample}.
}
