% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{dateSetCheck}
\alias{dateSetCheck}
\title{Check that dates match between \code{dateSet} and \code{dateTimeSet}}
\usage{
dateSetCheck(df)
}
\arguments{
\item{df}{A data frame to check}
}
\value{
ID's of rows that have mismatched dates.
}
\description{
\code{dateSetCheck} checks whether all the dates match between \code{dateSet} and \code{dateTimeSet}. If they don't, returns numeric indices of the rows that contain mismatches. This function also verifies that \code{df} is a data frame and that it has the columns \code{dateSet} and \code{dateTimeSet}.
}
