% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtil.R
\name{dbTable}
\alias{dbTable}
\title{Import an MFE database table, and (optionally) do a little bit of querying.}
\usage{
dbTable(table, lakeID = c(), depthClass = c(), fpath = dbdir, dbname = db)
}
\arguments{
\item{table}{The database table you want to import. \strong{Not} case-sensitive.}

\item{lakeID}{(optional) lakeID's to subset.}

\item{depthClass}{(optional) depthClasses to subset.}

\item{fpath}{By default, the \code{dbdir} variable, which is defined in the global environment (usually at the beginning of each analysis script)}

\item{dbname}{A character string containing the name of the .db file you want to access. By default, the globally-defined \code{db} variable is used.}
}
\value{
The chosen database table
}
\description{
Import an MFE database table, and (optionally) do a little bit of querying.
}
