% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtil.R
\name{dbTableList}
\alias{dbTableList}
\title{List names of the tables in the MFE database}
\usage{
dbTableList(fpath = dbdir, dbname = db)
}
\arguments{
\item{fpath}{By default, the \code{dbdir} variable, which is defined in the global environment (usually at the beginning of each analysis script)}

\item{dbname}{A character string containing the name of the .db file you want to access. By default, the globally-defined \code{db} variable is used.}
}
\value{
A vector of table names
}
\description{
List names of the tables in the MFE database
}
\examples{
dbdir <- here("currentDB")
db <- "MFEdb.db" # though note that it's best practice to specify a database version with date attached.
dbTableList(dbdir, db) # in this case, equivalent to dbTableList() or dbTableList(fpath = dbdir, dbname = db)
}
