% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtil.R
\name{dbTableSummary}
\alias{dbTableSummary}
\title{List column names in a MFE database table}
\usage{
dbTableSummary(
  table,
  cols = c("lakeID", "depthClass"),
  fpath = dbdir,
  dbname = db
)
}
\arguments{
\item{table}{The database table to explore. \strong{Not} case-sensitive.}

\item{cols}{A character vector of columns to summarize}

\item{fpath}{By default, the \code{dbdir} variable, which is defined in the global environment (usually at the beginning of each analysis script)}

\item{dbname}{A character string containing the name of the .db file you want to access. By default, the globally-defined \code{db} variable is used.}
}
\value{
A summary of the chosen columns
}
\description{
List column names in a MFE database table
}
