% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{dfCheck}
\alias{dfCheck}
\title{Check whether something is a data frame}
\usage{
dfCheck(obj)
}
\arguments{
\item{obj}{An object that you want to check for df-ness}
}
\value{
Throws an error if \code{obj} is not a data frame.
}
\description{
Check whether something is a data frame
}
