% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFEmetab.R
\name{fillHoles}
\alias{fillHoles}
\title{Fill Holes
Linearly interpolate values for strings (up to specified length) of missing data. Created by CTS, 31 July 2009.}
\usage{
fillHoles(dataIn, maxLength, timeStep)
}
\arguments{
\item{timeStep}{The time step of the data.}

\item{dataIn:}{A data frame with two columns, first is "datetime", second is data}

\item{maxLength:}{Maximum length of NA string that you are willing to interpolate across. NOTE that this is given in minutes.}
}
\value{
A data frame with same format as dataIn, but with values interpolated.
}
\description{
Fill Holes
Linearly interpolate values for strings (up to specified length) of missing data. Created by CTS, 31 July 2009.
}
