% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFEmetab.R
\name{findNotDupRows}
\alias{findNotDupRows}
\title{Find Not Dup Rows
Function to find duplicate datetime stamps. Returns the indices of rows where the datetime is NOT a duplicate of the datetime in a previous row.}
\usage{
findNotDupRows(dataIn)
}
\arguments{
\item{dataIn}{a data frame}
}
\value{
A vector, indices of rows where the datetime is NOT a duplicat of the datetime in a previous row.
}
\description{
Find Not Dup Rows
Function to find duplicate datetime stamps. Returns the indices of rows where the datetime is NOT a duplicate of the datetime in a previous row.
}
