% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFEmetab.R
\name{floorMins}
\alias{floorMins}
\title{Floor Mins
Round all time down to nearest timeStep (e.g. if timeStep is 5, round 00:07 to 00:05)}
\usage{
floorMins(dataIn, timeStep)
}
\arguments{
\item{dataIn}{A data frame}

\item{timeStep}{Time interval between DO measurements, minutes; defaults to 10 minutes}
}
\value{
A vector
}
\description{
Floor Mins
Round all time down to nearest timeStep (e.g. if timeStep is 5, round 00:07 to 00:05)
}
