% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{getNearPairsToCheck}
\alias{getNearPairsToCheck}
\title{Get Near Pairs to Check
This function picks out the sampleID's for pairs where the reps were close together (as per the user-defined threshold for distance in the checkDistance() function), but whose average value falls OOB. For each of those sampleID's, we will wind back the clock and take a look at each of the individual reps.}
\usage{
getNearPairsToCheck(avgsAndSingles)
}
\arguments{
\item{avgsAndSingles}{A data frame containing averaged near pairs and singletons.}
}
\value{
A vector of sampleID's that belong to averaged near-pairs (rather than singletons)
}
\description{
Get Near Pairs to Check
This function picks out the sampleID's for pairs where the reps were close together (as per the user-defined threshold for distance in the checkDistance() function), but whose average value falls OOB. For each of those sampleID's, we will wind back the clock and take a look at each of the individual reps.
}
