% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{lakeSiteCheck}
\alias{lakeSiteCheck}
\title{Lake Site Check
Given a lake, examine cumulative number of averages/singles at each site*depthClass combo over the years. This is just a convenience function I wrote to create the same ggplot over and over again without retyping the same code. (An additional function for checking QC codes. This was mostly useful to KG in DOC_QA_QC.R in Nov. 2020. Probably not for future use, but maybe!)}
\usage{
lakeSiteCheck(data, lake, alphaVal = 0.5)
}
\arguments{
\item{data}{Previous DOC data to examine.}

\item{lake}{LakeID to investigate (can only take one at a time, as a character string.)}

\item{alphaVal}{Transparency value for geom_line and geom_point. Defaults to 0.5.}
}
\value{
Creates a plot showing the cumulative sum of observations from each lake/depthClass/year. Returns the ggplot object.
}
\description{
Lake Site Check
Given a lake, examine cumulative number of averages/singles at each site*depthClass combo over the years. This is just a convenience function I wrote to create the same ggplot over and over again without retyping the same code. (An additional function for checking QC codes. This was mostly useful to KG in DOC_QA_QC.R in Nov. 2020. Probably not for future use, but maybe!)
}
