% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtil.R
\name{lookForDB}
\alias{lookForDB}
\title{Throw an error if trying to connect to a database file that doesn't exist. Avoids creating an 0B database file, which is the default behavior for dbConnect() otherwise.}
\usage{
lookForDB(path, db)
}
\arguments{
\item{path}{Directory where the database is stored}

\item{db}{File name of the database file you're trying to connect to}
}
\value{
If the database file doesn't exist, throws an error.
}
\description{
Throw an error if trying to connect to a database file that doesn't exist. Avoids creating an 0B database file, which is the default behavior for dbConnect() otherwise.
}
