% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFEmetab.R
\name{metabLoss_v4}
\alias{metabLoss_v4}
\title{Metab Loss, Version 4
Estimating Metabolism from DO data. This function calculates the likelihoods; use in conjunction with an optimization routine to minimize NLL. Use this for one day's data at a time. Created by CTS, 29 January 2009. Version 4 created August 2009.}
\usage{
metabLoss_v4(parsIn, dataIn)
}
\arguments{
\item{parsIn}{A vector of parameters. First element is \code{iota} (primary productivity per unit of PAR; units are (mg L-1 timeStep-1) / (mmol m-2 s-1)). Second element is \code{rho} (nighttime respiration; units are (mg L-1 (timeStep)-1)). Third element is DOInit (initial DOHat value; units are mg L-1). TimeStep is the number of minutes between observations. \code{parsIn} is given in log units.}

\item{dataIn}{A data frame, which includes the following columns: DOObs: DO (mg/L); DOSat: DO at saturation (mg/L); irr: irradiance (PAR) (mmol m-2 s-1) >>>> note that GLEON data may actually be in umol; check; kO2: piston velocity for O2, m (timeStep)-1; zMix: depth of the mixed layer (m); fluxDummy: 0 if zMix is above DO sensor depth (prevents atm flux); 1 otherwise.}
}
\value{
NLL, which is a scalar (I think? -KG)
}
\description{
Metab Loss, Version 4
Estimating Metabolism from DO data. This function calculates the likelihoods; use in conjunction with an optimization routine to minimize NLL. Use this for one day's data at a time. Created by CTS, 29 January 2009. Version 4 created August 2009.
}
