% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFEmetab.R
\name{metabPredix_v4}
\alias{metabPredix_v4}
\title{Metab Predix Version 4
Return DOHat, residuals, atmFlux, and NLL given par estimates. Created by CTS, 18 August 2009.}
\usage{
metabPredix_v4(parsIn, dataIn)
}
\arguments{
\item{parsIn}{A vector of parameters.}

\item{dataIn}{A data frame}
}
\value{
A data frame
}
\description{
Metab Predix Version 4
Return DOHat, residuals, atmFlux, and NLL given par estimates. Created by CTS, 18 August 2009.
}
