% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFEmetab.R
\name{mfeMetab}
\alias{mfeMetab}
\title{mfeMetab
Functionality for generating lake metabolism estimates from MFE database. Putting Solomon metabolism code into a function and adding some options for different K models. Also allows for pulling input data directly from MFE sensor database. WARNING! This function makes some modifications to your R workspace, which isn't best practice for R package functions. I don't have time to fix it now. It sets the system time zone, and it also saves your workspace to an .Rdata file. Use caution. -KG
SEJ with code from CTS, JAZ, CRO, and CJT}
\usage{
mfeMetab(
  lakeID,
  minDate,
  maxDate,
  outName,
  dirDump,
  maxZMix = 8,
  k = "cole&caraco",
  fluxDummyToggle = TRUE,
  bootstrap = "no",
  lat = 46.16,
  elev = 535,
  windHeight = 2,
  timeStep = 10,
  sensorDepth = 0.7
)
}
\arguments{
\item{lakeID}{Lake to estimate metabolism for}

\item{minDate}{First date to estimate metabolism for}

\item{maxDate}{Last date to estimate metabolism for}

\item{outName}{Text to use in lableing outputs, e.g. 'Acton2008'. Character. No default.}

\item{dirDump}{Directory where outputs should be dumped, e.g. 'C:/GLEON/Acton/Results'. No default value.}

\item{maxZMix}{Back up zmix if temp profile data is unavailable; defaults to 4, but should pick a value that makes sense.}

\item{k}{Model for piston velocity - options are: "cole&caraco", a constant k600 specified by the desired value, or "read" (coming soon)}

\item{fluxDummyToggle}{Use convention where atmospheric exchange is shut off during microstratification conditions.}

\item{lat}{Latitude of lake, decimal degrees, north positive; defaults to UNDERC (46.15)}

\item{elev}{Elevation of lake surface, m above sea level; defaults to UNDERC (518)}

\item{windHeight}{Height above lake surface at which wind speed is measured, m; defaults to UNDERC (2)}

\item{timeStep}{Time interval between DO measurements, minutes; defaults to 10 minutes}

\item{sensorDepth}{Depth of DO sensor, m; defaults to UNDERC (0.7?)}

\item{bootstrapping}{\code{yes} or \code{no}. Warning! Will take a while to run if \code{yes}. Defaults to \code{no}.}
}
\description{
mfeMetab
Functionality for generating lake metabolism estimates from MFE database. Putting Solomon metabolism code into a function and adding some options for different K models. Also allows for pulling input data directly from MFE sensor database. WARNING! This function makes some modifications to your R workspace, which isn't best practice for R package functions. I don't have time to fix it now. It sets the system time zone, and it also saves your workspace to an .Rdata file. Use caution. -KG
SEJ with code from CTS, JAZ, CRO, and CJT
}
