% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{named_group_split}
\alias{named_group_split}
\title{Named Group Split
Function written by Romain Francois, here: https://github.com/tidyverse/dplyr/issues/4223. Splits list into groups using group_split, but then names the list items according to a designation of your choice.}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{A data frame or tibble}

\item{...}{Variables to group by. Use like any other tidyverse function.}
}
\value{
A named list, where the names come from the groups used to split the original data frame/tibble.
}
\description{
Named Group Split
Function written by Romain Francois, here: https://github.com/tidyverse/dplyr/issues/4223. Splits list into groups using group_split, but then names the list items according to a designation of your choice.
}
\examples{
#mtcars \%>\% named_group_split(cyl)
}
