% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfuns.R
\name{parseSampleID}
\alias{parseSampleID}
\title{Parse sampleID's. KG NOTE: The <<- assignment here is not good practice; should assign to variables within the function and then have an object get returned. But I don't want to mess up existing workflows, so I'm leaving it like this for now.}
\usage{
parseSampleID(table, part, fish = FALSE)
}
\arguments{
\item{table}{A data frame}

\item{part}{Which part of the sampleID do you want to retrieve?}

\item{fish}{TRUE/FALSE is this a fish sampleID?}
}
\value{
Parsed sampleID portion.
}
\description{
Parse sampleID's. KG NOTE: The <<- assignment here is not good practice; should assign to variables within the function and then have an object get returned. But I don't want to mess up existing workflows, so I'm leaving it like this for now.
}
