% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{processAvgsAndSingles}
\alias{processAvgsAndSingles}
\title{Process Averages and Singletons
Process the avgsAndSingles data frame (averaged pairs and singletons) and assign them QC codes. This function does not deal with non-singleton rows with OOB == T--that's already been dealt with in getNearPairsToCheck}
\usage{
processAvgsAndSingles(avgsAndSingles)
}
\arguments{
\item{avgsAndSingles}{A data frame containing averaged near pairs and singletons.}
}
\value{
A data frame with QC codes applied.
}
\description{
Process Averages and Singletons
Process the avgsAndSingles data frame (averaged pairs and singletons) and assign them QC codes. This function does not deal with non-singleton rows with OOB == T--that's already been dealt with in getNearPairsToCheck
}
