% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{processCheckedNearPairs}
\alias{processCheckedNearPairs}
\title{Process Checked Near Pairs
Assign QC codes to near pairs whose average fell OOB.}
\usage{
processCheckedNearPairs(checkedDat)
}
\arguments{
\item{checkedDat}{A data frame containing the individual replicates for each sampleID that had a near pair whose average fell OOB. Each rep in this df has already been checked for OOB. Now we have to assign QCcodes.}
}
\value{
A data frame with QC codes applied.
}
\description{
Process Checked Near Pairs
Assign QC codes to near pairs whose average fell OOB.
}
