% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{processFarPairs}
\alias{processFarPairs}
\title{Process Far Pairs
Function to process far-apart pairs. Once we've flagged pairs of replicates as being far apart (as per the user-defined threshold specified in checkDistance()), we need to determine whether one replicate, both replicates, or neither replicate, is out of bounds. Then we flag accordingly.}
\usage{
processFarPairs(farPairs)
}
\arguments{
\item{farPairs}{A data frame (output from checkDistance())}
}
\value{
Data frame with OOB classifications for each replicate.
}
\description{
Process Far Pairs
Function to process far-apart pairs. Once we've flagged pairs of replicates as being far apart (as per the user-defined threshold specified in checkDistance()), we need to determine whether one replicate, both replicates, or neither replicate, is out of bounds. Then we flag accordingly.
}
