% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_QA_QC_funs.R
\name{processNewData}
\alias{processNewData}
\title{Process New Data (main function)
This is the main function, a wrapper that calls all the preceding functions. The user won't really interact with the functions defined above, just with this one.}
\usage{
processNewData(newData, oldData = newData[FALSE, ])
}
\arguments{
\item{newData}{Newly-collected DOC data to be processed (a data frame)}

\item{oldData}{Previous DOC data, typically from the WATER_CHEM database table. If no other argument is supplied, the default is a blank version of newData (same column names, but no content).}
}
\value{
A data frame in the appropriate format to be put in the database, with QC codes applied.
}
\description{
Process New Data (main function)
This is the main function, a wrapper that calls all the preceding functions. The user won't really interact with the functions defined above, just with this one.
}
