% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfuns.R
\name{recreateSampleIDs}
\alias{recreateSampleIDs}
\title{Attempt to recreate sampleID's from component parts. KG note: I'm not super clear on how this function works. It gets the component parts from the sampleID itself, rather than taking them from other columns, so under what circumstances would the sampleID not be able to be re-created?}
\usage{
recreateSampleIDs(table, fish = FALSE)
}
\arguments{
\item{table}{A data frame}

\item{fish}{TRUE/FALSE is this a fish sampleID?}
}
\value{
Messages telling you whether the sampleID's can be re-created.
}
\description{
Attempt to recreate sampleID's from component parts. KG note: I'm not super clear on how this function works. It gets the component parts from the sampleID itself, rather than taking them from other columns, so under what circumstances would the sampleID not be able to be re-created?
}
