% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{sameTime}
\alias{sameTime}
\title{Check for rows where \code{dateTimeSet} and \code{dateTimeSample} are the same}
\usage{
sameTime(df, excludeZero = F)
}
\arguments{
\item{df}{A data frame to check}

\item{excludeZero}{Logical}
}
\value{
ID's of rows that have mismatched dates.
}
\description{
\code{sameTime} checks whether there are any rows where \code{dateTimeSet} is equal to \code{dateTimeSample}. The function also verifies that \code{df} is a data frame and has \code{dateTimeSet} and \code{dateTimeSample} columns.
}
