% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensordbTable.R
\name{sensordbTableList}
\alias{sensordbTableList}
\title{Sensor DB Table List
Listing table names in the MFE sensor database. Created by JAZ, 2015-11-30}
\usage{
sensordbTableList(fpath = sensor_dbdir, dbname = sensor_db)
}
\arguments{
\item{fpath}{File path of database file location. Defaults to whatever you set as \code{sensor_dbdir} in your script.}

\item{dbname}{Name of the database file. Default is whatever you've set to the variable \code{sensor_db} in}
}
\value{
Vector of database table names
}
\description{
Sensor DB Table List
Listing table names in the MFE sensor database. Created by JAZ, 2015-11-30
}
