% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{timeTravel}
\alias{timeTravel}
\title{Check that sample-set time spans are logical}
\usage{
timeTravel(df)
}
\arguments{
\item{df}{A data frame to check}
}
\value{
ID's of rows that have mismatched dates.
}
\description{
\code{timeTravel} checks whether there are any rows where \code{dateTimeSet} is later than \code{dateTimeSample}, which wouldn't make sense. The function also verifies that \code{df} is a data frame and has \code{dateTimeSet} and \code{dateTimeSample} columns.
}
